requirements.txt
pillow~=8.4.0
keras~=2.6.0
tensorflow~=2.6.0
numpy~=1.19.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1